Store images created with `filePath` on girder
Now, if a user creates a stem image with a `filePath` parameter:

1. The user must be an administrator (no one else should be able to read
files on the server).
2. The file is imported to the girder server and put in a private
folder called `stem_images` in the root folder of the user who made it.
If the folder already exists, it will be re-used.
3. The file is put in an item called `_imported_images`. The names of
the files are all just indices (starting with 1). If 1000 files exist,
an error will occur when trying to import another one.
4. When the stem image is deleted, if it was an imported file, the
imported file is deleted as well.

These changes allow us to enforce all stem images to have a girder
`fileId`.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>