Use generator func for getting dataset in chunks
This allows a max chunk size to be set for sending the arrays
over the network, and calculates a read size based upon that.
However, it will always read at least one array, even if the
max chunk size is exceeded.

This function will hopefully be re-usable by the msgpack format.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>