Send all frames in chunks through msgpack as well
Since this is probably the most amount of data that gets sent
over at once, it is important to send it in chunks.

Unfortunately, it is also difficult to send in chunks since it is
an array of variable length arrays.

This does work, though. And it can be unpacked as follows:
```
res = gc.get('/stem_images/%s/frames' % stem_image_id, jsonResp=False)

unpacker = msgpack.Unpacker(raw=False)
unpacker.feed(res.content)
frame_events = []
for val in unpacker:
    frame_events += val
```

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>