Allow an index to be used instead of name
This checks to see if the name is an integer. If it is, then
it assumes that it is an index, and it uses it as an index
instead of as a name.

It also has error checking to make sure the index does not
exceed the length of the dataset.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>