Allow not-logged-in users to use all GET methods
This allows all users who are not logged in to use every
GET method. Previously, it would say that "you must be logged
in" in order to use most of the GET methods. But we want
users who are not logged in to be able to access public data,
so allow not-logged-in users to use the GET methods.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>