Put imported stem images in public/private folders
Previously, all imported stem images (that is, those created with
the "filePath" on the local filesystem) were placed in a private folder.
But this meant that imported stem image files that were marked as
`public` would not be readable by others.

To fix this, if a created stem image is marked as `public` and it
is created with a "filePath" on the local filesystem, put the
stem image file in a public folder. Otherwise, default to putting
the stem image in a private folder.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>